Garllery-API/pyproject.toml
[tool.poetry]
name = "alpha-gallery"
version = "0.1.0"
description = ""
authors = ["Basel Mousa Atalla <baselatalla10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-cors-headers = "^3.7.0"
Pillow = "^8.3.1"
psycopg2 = "^2.9.1"
django-rest-auth = "^0.9.5"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]
Django = "^3.2.6"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
